Version 2.0 - Complete Recode The whole project/plugin has been recoded from scratch, as the old code was a one big mess.
This means plugin is now optimized better and future development will be much easier, as it is now very flexible and easy to work with.

! IMPORTANT
Old configurations unfortunately won't work, as the plugin now handles files a bit differently.
I recommend saving the old Marriage folder and letting a new one create, then it is pretty much just copy and paste of the old values.
!! BUT DON'T COPY THE WHOLE CONFIG SECTIONS !!
Please, give the config files a little look and paste ONLY the values, to prevent any unwanted errors.